+++ src/frformat/geo_enum_format.py
@@ -1,30 +1,16 @@
 from enum import Enum, auto
-from functools import total_ordering
 from typing import Dict, FrozenSet, Type
 
 from frformat import CustomStrFormat, Metadata
 from frformat.common import normalize_value
 from frformat.options import Options
-from frformat.versioned_set import Version
 
 
-@total_ordering
-class Millesime(Enum, Version):
+class Millesime(Enum):
     M2023 = auto()
     M2024 = auto()
 
-    def __eq__(self, other) -> bool:
-        return self.value == other.value
-
-    def __lt__(self, other) -> bool:
-        return self.value < other.value
-
-    def get_id(self) -> str:
-        return str(self.value)
-
-    @classmethod
-    def is_sorted(cls) -> bool:
-        return True
+    LATEST = M2024
 
 
 def new(
